Renamed "check" package to "checks"
