+++ proptool/checks/base/check.py
