+++ proptool/checks/brackets.py
