+++ proptool/checks/dangling_keys.py
