+++ proptool/checks/empty_translations.py
