+++ proptool/checks/formatting_values.py
