+++ proptool/checks/key_format.py
