+++ proptool/checks/missing_translation.py
