+++ proptool/checks/punctuation.py
