+++ proptool/checks/starts_with_the_same_case.py
