+++ proptool/checks/trailing_white_chars.py
