+++ proptool/checks/white_chars_before_linefeed.py
